Force rename git MapView in git
